+++ src/pages/Home.tsx
@@ -7,10 +7,12 @@ import WhyUs from '../sections/WhyUs';
 import HelpCta from '../sections/HelpCta';
 import Testimonials from '../sections/Testimonials';
 import Contact from '../sections/Contact';
+import CursorFlow from '../components/CursorFlow';
 
 export default function Home() {
   return (
     <>
+      <CursorFlow />
       <Hero />
       <Features />
       <About />